requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
black==23.1.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
chromedriver-py==110.0.5481.30
click==8.1.3
coverage==7.3.2
dataclasses==0.6
dataclasses-json==0.5.7
exceptiongroup==1.1.0
fake-headers==1.0.2
fake-useragent==1.2.1
frozenlist==1.3.3
geckodriver-autoinstaller==0.1.0
gender-guesser==0.4.0
h11==0.14.0
html5lib==1.1
hypothesis==6.68.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
isort==5.12.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mock==5.0.1
multidict==6.0.4
mypy-extensions==0.4.3
nameparser==1.1.2
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.1
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2022.7.1
rdflib==6.2.0
requests==2.28.2
responses==0.22.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SPARQLWrapper==2.0.0
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
types-toml==0.10.8.4
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
useragent==0.1.1
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.2
